move title functions to utils. separate out get and update titleElem
